Refactored DataSourceForm.tsx to styled. Decomposed NamespaceAutocomplete.tsx from DataSourceForm.tsx
