+++ odd-platform-ui/src/components/Management/DataSourcesList/DataSourceForm/DataSourceFormStyles.ts
@@ -1,0 +1,5 @@
+import { styled } from '@mui/material/styles';
+
+export const Asterisk = styled('span')(({ theme }) => ({
+  color: theme.palette.warning.main,
+}));
