+++ odd-platform-ui/src/components/Management/DataSourcesList/DataSourceForm/NamespaceAutocomplete/NamespaceAutocompleteContainer.tsx
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import { fetchNamespaceList } from 'redux/thunks';
+import NamespaceAutocomplete from './NamespaceAutocomplete';
+
+const mapStateToProps = () => ({});
+
+const mapDispatchToProps = {
+  searchNamespace: fetchNamespaceList,
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(NamespaceAutocomplete);
