+++ odd-platform-ui/src/components/Management/DataSourcesList/DataSourceItem/DataSourceItem.tsx
@@ -10,7 +10,7 @@ import ConfirmationDialog from 'components/shared/ConfirmationDialog/Confirmatio
 import BooleanFormatted from 'components/shared/BooleanFormatted/BooleanFormatted';
 import EditIcon from 'components/shared/Icons/EditIcon';
 import DeleteIcon from 'components/shared/Icons/DeleteIcon';
-import DataSourceFormDialogContainer from 'components/Management/DataSourcesList/DataSourceFormDialog/DataSourceFormDialogContainer';
+import DataSourceFormDialogContainer from 'components/Management/DataSourcesList/DataSourceForm/DataSourceFormContainer';
 import AppButton from 'components/shared/AppButton/AppButton';
 import * as S from './DataSourceItemStyles';
